feat: add some very basic tree shaking
There is not much we can do to optimise the actual search logic, but
what we can do it shake the rule down into its most optimal for do
reduce object accesses. As this is where the biggest gains can be
obtained (each access with HashMaps is around 100ns). This code does not
do a full tree shake but handles some of the common cases.

This commit also improves the parser a little bit and fixes some bugs.