feat: add in the matrix boolean type
With rules that make use of heavy whitelisting we end up spending a lot
of time looking up the same fields, there is no easy way around this due
to the design of the solver. Although by adding a new expression type
`matrix` and making use of an array cache it is possible to only perform
the lookup once per field. This can only be applied to certain boolean
logic within Tau. There are further improvements we can do here later
down the line.