feat: allow traits to be made thread safe via flag
This should probably be default but I don't want to break the
implementation if anyone is doing unthreadsafe stuff under the hood. I
also feel that changing a traits attributes with a feature flag is
pretty hacky. We should probably do this in a cleaner way like rust does
i.e. std::{sync, unsync}, but that is more work and mutal exclusivity is
fine for now...